Code restructure
- Flatten the file hierachy
- Remove plugin
- Remove Python binding

Python plugin provided by the users should be forbidden as the
software is optimized for performance.
