+++ README.md
@@ -7,7 +7,7 @@ module load gcc/9.3.0
 conda create -n tomcat-live python==3.7.10
 
 conda activate tomcat-live
-conda install -c conda-forge cmake cppzmq eigen boost fftw libastra tbb-devel nlohmann_json spdlog pybind11
+conda install -c conda-forge cmake cppzmq eigen boost fftw libastra tbb-devel nlohmann_json spdlog
 
 git clone --recursive <repo>
 
@@ -16,7 +16,7 @@ git clone --recursive <repo>
 # the Python bindings 
 mkdir build && cd build
 cmake .. -DCMAKE_PREFIX_PATH=${CONDA_PREFIX:-"$(dirname $(which conda))/../"} \
-         -DBUILD_PYTHON=ON -DBUILD_MONITOR=ON 
+         -DBUILD_MONITOR=ON 
 make -j12 && make install
 
 # On the graphics workstation x02da-gws-3
