--- slicerecon/include/slicerecon/slicerecon.hpp
@@ -1,8 +1,0 @@
-#include "reconstruction/helpers.hpp"
-#include "reconstruction/reconstructor.hpp"
-#include "reconstruction/projection_processor.hpp"
-#include "reconstruction/utils.hpp"
-#include "plugin.hpp"
-#include "projection_server.hpp"
-#include "visualization_server.hpp"
-#include "data_types.hpp"
