+++ slicerecon/include/slicerecon/solver.hpp
@@ -18,7 +18,7 @@
 #include "astra/ParallelVecProjectionGeometry3D.h"
 #include "astra/VolumeGeometry3D.h"
 
-#include "slicerecon/data_types.hpp"
+#include "data_types.hpp"
 
 
 namespace slicerecon {
