+++ tomopackets/CMakeLists.txt
@@ -14,14 +14,3 @@ if (BUILD_TEST)
     target_link_libraries(test_server PRIVATE cppzmq pthread Boost::boost)
     target_compile_options(test_server PRIVATE -Wall -Wextra -Wfatal-errors)
 endif()
-
-if (BUILD_PYTHON)
-    set(BINDING_NAME "tomop")
-    pybind11_add_module(${BINDING_NAME} "python/tomop.cpp")
-    target_include_directories(${BINDING_NAME} PRIVATE "${CMAKE_CURRENT_LIST_DIR}/include")
-    target_link_libraries(${BINDING_NAME} PRIVATE cppzmq pthread Boost::boost)
-    target_compile_options(${BINDING_NAME} PRIVATE -Wall -Wextra -Wfatal-errors)
-
-    install(TARGETS ${BINDING_NAME}
-            LIBRARY DESTINATION ${Python3_SITELIB})
-endif()
